requirements.txt
Pillow
fastmcp
pydantic_ai
pydantic-ai[openai]
python-dotenv
logfire